Delete base classes
